+++ src/frontend/apps/impress/src/components/modal/ButtonCloseModal.tsx
@@ -1,23 +1,32 @@
-import { Button, type ButtonProps } from '@gouvfr-lasuite/cunningham-react';
-import React from 'react';
+import {
+  Button,
+  ButtonElement,
+  type ButtonProps,
+} from '@gouvfr-lasuite/cunningham-react';
+import React, { forwardRef } from 'react';
 
 import { Icon } from '@/components';
 
-export const ButtonCloseModal = (props: ButtonProps) => {
-  return (
-    <Button
-      type="button"
-      size="small"
-      color="brand"
-      variant="tertiary"
-      icon={
-        <Icon
-          $withThemeInherited
-          iconName="close"
-          className="material-icons-filled"
-        />
-      }
-      {...props}
-    />
-  );
-};
+export const ButtonCloseModal = forwardRef<ButtonElement, ButtonProps>(
+  (props, ref) => {
+    return (
+      <Button
+        ref={ref}
+        type="button"
+        size="small"
+        color="brand"
+        variant="tertiary"
+        icon={
+          <Icon
+            $withThemeInherited
+            iconName="close"
+            className="material-icons-filled"
+          />
+        }
+        {...props}
+      />
+    );
+  },
+);
+
+ButtonCloseModal.displayName = 'ButtonCloseModal';